# lib/api-spec/openapi.yaml
openapi: 3.1.0
info:
  title: Api
  version: 0.1.0
  description: Chop Plan API - prepaid lunch subscription service for Lagos restaurants
servers:
  - url: /api
    description: Base API path
...
paths:
...
  /user/profile:
    get:
      operationId: getUserProfile
      tags: [user]
      summary: Get logged-in user profile
      responses:
        "200":
          description: User profile
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/UserProfile"
    patch:
      operationId: updateUserProfile
      tags: [user]
      summary: Update user profile
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/UserProfileUpdate"
      responses:
        "200":
          description: Updated
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/UserProfile"
  /user/subscriptions:
    get:
      operationId: listUserSubscriptions
      tags: [user]
      summary: Get the user's active subscriptions
      responses:
        "200":
          description: Subscriptions
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/UserSubscription"
    post:
      operationId: createSubscription
      tags: [user]
      summary: Subscribe to a vendor plan
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/SubscriptionInput"
      responses:
        "201":
          description: Created
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/UserSubscription"
  /user/subscriptions/{subscriptionId}:
    delete:
      operationId: cancelSubscription
      tags: [user]
      summary: Cancel a subscription
      parameters:
        - name: subscriptionId
          in: path
          required: true
          schema:
            type: integer
      responses:
        "200":
          description: Cancelled
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/SuccessResponse"
  /user/subscriptions/{subscriptionId}/schedule:
    get:
      operationId: getUserSubscriptionSchedule
      tags: [user]
      summary: Get the daily pickup schedule for a subscription
      parameters:
        - name: subscriptionId
          in: path
          required: true
          schema:
            type: integer
      responses:
        "200":
          description: Schedule days
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/ScheduleDay"
        "404":
          description: Not found
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
  /user/subscriptions/{subscriptionId}/schedule/{dayId}/confirm:
    post:
      operationId: confirmPickup
      tags: [user]
      summary: Confirm a day's pickup was received
      parameters:
        - name: subscriptionId
          in: path
          required: true
          schema:
            type: integer
        - name: dayId
          in: path
          required: true
          schema:
            type: integer
      responses:
        "200":
          description: Confirmed
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ScheduleDay"
        "400":
          description: Cannot confirm this day
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
        "404":
          description: Not found
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"

  # ── VENDOR dashboard ──────────────────────────────────────────────────────
...
components:
  schemas:
...
    SuccessResponse:
      type: object
      properties:
        success:
          type: boolean
        message:
          type: string
      required: [success]
    ErrorResponse:
      type: object
      properties:
        error:
          type: string
      required: [error]

    # ── Auth ──────────────────────────────────────────────────────────────
...
    UserProfile:
      type: object
      required: [id, name, email, phone, area]
      properties:
        id:
          type: integer
        name:
          type: string
        email:
          type: string
        phone:
          type: string
        area:
          type: string
    UserProfileUpdate:
      type: object
      properties:
        name:
          type: string
        phone:
          type: string
        area:
          type: string
    UserSubscription:
      type: object
      required: [id, vendorId, vendorName, planName, daysPerMonth, freeDays, priceNaira, startDate, status]
      properties:
        id:
          type: integer
        vendorId:
          type: integer
        vendorName:
          type: string
        vendorCoverImage:
          type: ["string", "null"]
        planName:
          type: string
        daysPerMonth:
          type: integer
        freeDays:
          type: integer
        priceNaira:
          type: integer
        startDate:
          type: string
          format: date
        status:
          type: string
          enum: [active, paused, cancelled]
    SubscriptionInput:
      type: object
      required: [vendorId, planId]
      properties:
        vendorId:
          type: integer
        planId:
          type: integer

    # ── Vendor dashboard ──────────────────────────────────────────────────
...
    ScheduleDay:
      type: object
      required: [id, dayNumber, scheduledDate, status, confirmedAt]
      properties:
        id:
          type: integer
        dayNumber:
          type: integer
        scheduledDate:
          type: string
          format: date
        status:
          type: string
          enum: [pending, confirmed]
        confirmedAt:
          type: ["string", "null"]
          format: date-time
...