# lib/api-zod/src/generated/types/scheduleDay.ts
/**
 * Generated by orval v8.20.0 🍺
 * Do not edit manually.
 * Api
 * Chop Plan API - prepaid lunch subscription service for Lagos restaurants
 * OpenAPI spec version: 0.1.0
 */
import type { ScheduleDayStatus } from './scheduleDayStatus';

export interface ScheduleDay {
  id: number;
  dayNumber: number;
  scheduledDate: Date;
  status: ScheduleDayStatus;
  /** @nullable */
  confirmedAt: Date | null;
}